chore: Bring in the updated environment/docker infrastructure
This should have been in the 1.9.0 release but got forgetten about

+++ babs/differential/resources/shell/server-info.sh
@@ -3,7 +3,9 @@ server_info() {
     # $2 = usual port
     # $3 = option url suffix
     # Trap SIGINT/SIGTERM and kill *this specific singularity exec*
-    trap "echo '⚠️ Stopping $1 (PID=$PID)...'; kill $PID; wait $PID 2>/dev/null; exit 0" SIGINT SIGTERM
+    if [ -z "$run_tmux" ]; then 
+	trap "echo '⚠️ Stopping $1 (PID=$PID)...'; kill $PID; wait $PID 2>/dev/null; exit 0" SIGINT SIGTERM
+    fi
     echo "✅ Starting $1 server..."
     # Wait until the port responds
     for i in {1..120}; do
@@ -17,7 +19,7 @@ server_info() {
     fi
     ## Produce the informative message
     echo "✅ Point your web browser to http://$(hostname).$(dnsdomainname):$PORT$3"
-    command -v doitclient >/dev/null 2>&1 && doitclient www "http://$(hostname).$(dnsdomainname):$PORT$3" || true
+    command -v doitclient >/dev/null 2>&1 && timeout 10s doitclient www "http://$(hostname).$(dnsdomainname):$PORT$3" || true
 
     if [ -n "${SLURM_JOB_ID}" ]; then
         cat 1>&2 <<END
@@ -32,7 +34,7 @@ END
     fi
 
     if [ -z "${SLURM_JOB_ID}" ]; then
-        echo "⚠️ When done using using the server, quit (Ctrl+C) this process"
+        [ -z "$run_tmux" ] && echo "⚠️ When done using using the server, quit (Ctrl+C) this process" || echo "⚠️ When done using using the server, close the tmux window (Ctrl+C in it)"
     else
         cat 1>&2 <<END
 ⚠️ When done using using the app, issue the following command on the ${login_node} node:
